feat: add ClusterOriginIssuer
When using this controller in single-tenant clusters it becomes
desirable for many users to have a single cluster-scoped issuer rather
than each namespace duplicating the issuers and secrets.

This changeset introduces a "cluster resource namespace", which defaults
to the same namespace as the controller (eg, "origin-ca-issuer"). A
ClusterOriginIssuer resource type has been added which resolves
SecretKeySelectors in this namespace. Other than determing what
namespace to resolve the secret, the implementation is the same as the
existing OriginIssuer.

Fixes: #24